reworking ui build
