+++ scripts/run
@@ -93,7 +93,7 @@ fi
 /tmp/demo-app &
 
 cd "../server"
-go run cmd/build-assets/build-assets.go
+go run cmd/build-ui/main.go
 protoc --proto_path=. --plugin=protoc-gen-go=../build/protoc-gen-go --go_out=. --plugin=protoc-gen-twirp=../build/protoc-gen-twirp --twirp_out=. -Iproto/ $(find . -iname "*.proto")
 go build -tags development -o /tmp/kaja ./cmd/server
 /tmp/kaja
